# openapi.yaml
openapi: 3.0.0
info:
  title: PLLAY Enterprise Solution API
  version: 1.0.0
  description: API documentation for PLLAY Enterprise Solution
servers:
  - url: https://api.pllay.com/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /games:
    get:
      summary: Get list of games
      tags:
        - Games
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                        status:
                          type: string
                        popularity:
                          type: number
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer

    post:
      summary: Create a new game
      tags:
        - Games
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - apiEndpoint
              properties:
                name:
                  type: string
                description:
                  type: string
                apiEndpoint:
                  type: string
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  apiEndpoint:
                    type: string
                  status:
                    type: string
                  popularity:
                    type: number
  /games/{id}:
    get:
      summary: Get a specific game
      tags:
        - Games
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  description:
                    type: string
                  apiEndpoint:
                    type: string
                  status:
                    type: string
                  popularity:
                    type: number

    put:
      summary: Update a game
      tags:
        - Games
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                apiEndpoint:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Game updated successfully

    delete:
      summary: Delete a game
      tags:
        - Games
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game deleted successfully
...
  /admin/users:
    get:
      summary: Get all users (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        email:
                          type: string
                        role:
                          type: string
                        status:
                          type: string
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
...